eradicate the use of exceptions for request validation

+++ ca/src/ReqTrack/Domain.Core.UseCases/Users/DeleteUser/DeleteUserUseCase.cs
@@ -2,10 +2,8 @@
 using ReqTrack.Domain.Core.Exceptions;
 using ReqTrack.Domain.Core.Repositories;
 using ReqTrack.Domain.Core.Security;
-using ReqTrack.Domain.Core.UseCases.Boundary.Extensions;
 using ReqTrack.Domain.Core.UseCases.Boundary.Interfaces;
 using ReqTrack.Domain.Core.UseCases.Boundary.Responses;
-using ReqTrack.Domain.Core.UseCases.Exceptions;
 
 namespace ReqTrack.Domain.Core.UseCases.Users.DeleteUser
 {
@@ -25,7 +23,10 @@ namespace ReqTrack.Domain.Core.UseCases.Users.DeleteUser
         {
             try
             {
-                request.ValidateAndThrowOnInvalid();
+                if (!request.Validate(out var errors))
+                {
+                    return output.Accept(new ValidationErrorResponse(errors, "Invalid request."));
+                }
 
                 if (!_userRepository.DeleteUser(request.UserId))
                 {
@@ -36,10 +37,6 @@ namespace ReqTrack.Domain.Core.UseCases.Users.DeleteUser
                 {
                     Message = "User deleted successfully.",
                 });
-            }
-            catch (RequestValidationException e)
-            {
-                return output.Accept(new ValidationErrorResponse(e.ValidationErrors, $"Invalid request. {e.Message}"));
             }
             catch (EntityNotFoundException e)
             {
